Return a shared_ptr rather than a naked pointer
So far, the MS created was stored in a pointer internal to the class but it
was also returned to the client. One of the clients is the MsFactory class in
casa/code, which also passes the pointer to the requester. This can be rather
confusing in terms of who owns actually the pointer and should deallocate it.
A shread_ptr will hopefully help to keep the lifecycle of the pointer better
controlled.

A very basic unit test on the NewMSSimulator class has also been added.