Table iterator: Reuse a RefTable for each iteration.
This will reuse a RefTable created at construction time
to avoid paying the cost of a RefTable creation everytime
the next() method is called to advance iteration.
When next() is called then the RefTable is reset with
a call to RefTable::removeAllRow() and then the rows for
this iteration are added.
In a dataset with 10 rows per iteration the cost
goes down from ~80 usecs per next() call to ~13 usecs.